+++ test/reporters/human/common.test.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const common = require('../reporters/human/common')
+const common = require('../../../reporters/human/common')
 
 it('returns current time', () => {
   expect(common.now().valueOf()).toBeCloseTo(Date.now(), -2)
